Refactor CommandResult into its own module
