+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -9,7 +9,7 @@ import javafx.util.Pair;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.commandresult.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.MasterDeckParser;
 import seedu.address.logic.parser.exceptions.ParseException;
