+++ src/main/java/seedu/address/logic/commands/ExitCommand.java
@@ -1,5 +1,6 @@
 package seedu.address.logic.commands;
 
+import seedu.address.logic.commands.commandresult.CommandResult;
 import seedu.address.model.Model;
