+++ src/main/java/seedu/address/logic/commands/cardcommands/AddCommand.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 import java.util.Set;
 
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.commandresult.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.card.Answer;
