+++ src/main/java/seedu/address/logic/commands/deckcommands/AddDeckCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands.deckcommands;
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.commandresult.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.deck.Deck;
