+++ src/main/java/seedu/address/logic/commands/reviewcommands/MarkWrongCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands.reviewcommands;
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.commandresult.CommandResult;
 import seedu.address.model.Model;
 
 /**
